requirements.txt
#rodar ' pipreqs . --encoding=utf-8 ' na raiz do projeto para atualizar

altair==5.5.0
mysql-connector-python
pandas==2.3.3
plotly==6.5.0
python-dotenv==1.2.1
streamlit==1.51.0
